refac: move file struct to model.go
