+++ server/server.go
@@ -6,22 +6,11 @@ import (
 	"log"
 	"net/http"
 	"os"
-	"time"
 
 	"gorm.io/driver/mysql"
 	"gorm.io/gorm"
 	"github.com/joho/godotenv"
 )
-
-type File struct {
-	ID         int       `gorm:"primaryKey;autoIncrement"`
-	Name       string    `gorm:"type:varchar(255);not null"`
-	Path       string    `gorm:"type:varchar(255);not null"`
-	HashDigest string    `gorm:"type:varchar(256)"`
-	Content    string    `gorm:"type:text;not null"`
-	CreatedAt  time.Time `gorm:"type:datetime"`
-	UpdatedAt  time.Time `gorm:"type:datetime"`
-}
 
 // Simple function to ping and test if server is up or not
 func getPing(w http.ResponseWriter, r *http.Request) {
@@ -47,7 +36,7 @@ func main() {
 		dbUser, dbPassword, dbHost, dbPort, dbName)
 
 	// Connect to the database using GORM
-	db, err := gorm.Open(mysql.Open(dsn), &gorm.Config{})
+	_, err := gorm.Open(mysql.Open(dsn), &gorm.Config{})
 	if err != nil {
 		log.Fatalf("Failed to connect to the database: %v", err)
 	}
